+++ .gitignore
@@ -12,3 +12,4 @@ node_modules
 
 # Test coverage
 coverage/
+.nyc_output/
