+++ .travis.yml
@@ -13,5 +13,5 @@ before_script:
   - chmod +x ./cc-test-reporter
 
 script:
-  - npm run ci-test
+  - npm run test
   - if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then ./cc-test-reporter after-build --exit-code $TRAVIS_TEST_RESULT; fi
