+++ demo/test/webpack-demo-test.config.js
@@ -5,7 +5,7 @@ var path = require("path");
 var parentConfig = require("../webpack-demo.config");
 var nodeExternals = require("webpack-node-externals");
 
-var config = {
+var config = Object.assign(parentConfig, {
 
   // where does the import chain start
   entry: path.resolve(__dirname, "main.js"),
@@ -15,6 +15,18 @@ var config = {
 
   target: "node",
 
-};
+});
 
-module.exports = Object.assign(parentConfig, config);
+config.module.rules.push(
+  {
+    test: /\.js$/,
+    include: [
+      path.resolve(__dirname, ".."),
+      path.resolve(__dirname, "..", "..", "src"),
+    ],
+    loader: "istanbul-instrumenter-loader",
+    enforce: "post"
+  }
+);
+
+module.exports = config;
