+++ package.json
@@ -6,9 +6,7 @@
   "scripts": {
     "build": "webpack",
     "build-demo": "webpack --config ./demo/webpack-demo.config.js",
-    "test": "mocha-webpack --webpack-config ./test/webpack-test.config.js --recursive test",
-    "test-demo": "mocha-webpack --webpack-config ./demo/test/webpack-demo-test.config.js --recursive demo/test",
-    "ci-test": "istanbul cover mocha-webpack --report lcovonly -- --webpack-config ./test/webpack-test.config.js --recursive test",
+    "test": "nyc --reporter=lcov --reporter=text mocha-webpack --webpack-config ./demo/test/webpack-demo-test.config.js --recursive demo/test",
     "serve-demo": "webpack-dev-server --config ./demo/webpack-demo.config.js"
   },
   "repository": {
@@ -25,6 +23,9 @@
   },
   "homepage": "https://github.com/galarant/domless",
   "devDependencies": {
+    "babel-core": "6.25.0",
+    "babel-loader": "7.1.1",
+    "babel-preset-env": "1.6.0",
     "canvas-prebuilt": "1.4.0",
     "chai": "3.5.0",
     "esdoc": "0.5.2",
@@ -35,6 +36,7 @@
     "jsdom": "9.12.0",
     "mocha": "3.3.0",
     "mocha-webpack": "0.7.0",
+    "nyc": "11.0.3",
     "url-loader": "0.5.8",
     "webpack": "2.4.1",
     "webpack-chain": "3.3.0",
@@ -46,5 +48,14 @@
     "lodash": "4.17.4",
     "mocha-webpack": "0.7.0",
     "phaser": "2.6.2"
+  },
+  "nyc": {
+    "include": [
+      "demo/**/*.js",
+      "src/**/*.js"
+    ],
+    "instrument": false,
+    "report-dir": "./coverage",
+    "sourceMap": false
   }
 }
